requirements.txt
aiohttp==3.7.4
discord.py==1.7.1
PyNaCl==1.4.0
youtube_dl
youtube_search